Remove all old database files
